revert file move
